+++ src/handlers/teacher/__init__.py
@@ -1,12 +1,11 @@
 from aiogram import Router
 
-from .all_week import AllWeekHandler
 from .my_post import MyPostHandler
-from .lessons_by_day import LessonsByDaysHandler
 from .academy_time import AcademyTimeHandler
 from .olymp import get_olymp_router
 from .form import get_form_routers
 from src.middlewares import RoleAccessMiddleware
+from .schedule import ScheduleHandler
 
 
 def get_teacher_router() -> Router:
@@ -14,9 +13,8 @@ def get_teacher_router() -> Router:
     router = Router(name='teacher')
 
     routers = [
-        AllWeekHandler().get_router(),
         MyPostHandler().get_router(),
-        LessonsByDaysHandler().get_router(),
+        ScheduleHandler().get_router(),
         AcademyTimeHandler().get_router(),
         *get_olymp_router(),
         *get_form_routers()
